Calculate EDGE index

Calculate EDGE index

Usage

calculate_EDGE1(tree, table, sort.list = FALSE)

Arguments

tree

a phylo object.

table

a tibble or data frame with two columns, named "species" and "RL.cat"

sort.list

Logical. If TRUE, the EDGE list will be sorted from higher to lower values.

Value

a data frame with the input table and the EDGE metric (sensu Isaac et al., 2007).

References

Isaac, N.J., Turvey, S.T., Collen, B., Waterman, C. & Baillie, J.E. (2007) Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2: e296. https://doi.org/10.1371/journal.pone.0000296